Move CameraPosition to model
