+++ app/src/main/java/org/groundplatform/android/ui/map/gms/GoogleMapsFragment.kt
@@ -50,9 +50,9 @@ import org.groundplatform.android.model.imagery.LocalTileSource
 import org.groundplatform.android.model.imagery.RemoteMogTileSource
 import org.groundplatform.android.model.imagery.TileSource
 import org.groundplatform.android.model.map.Bounds
+import org.groundplatform.android.model.map.CameraPosition
 import org.groundplatform.android.persistence.remote.RemoteStorageManager
 import org.groundplatform.android.ui.common.AbstractFragment
-import org.groundplatform.android.ui.map.CameraPosition
 import org.groundplatform.android.ui.map.Feature
 import org.groundplatform.android.ui.map.MapFragment
 import org.groundplatform.android.ui.map.MapType
