+++ app/src/test/java/org/groundplatform/android/ui/datacollection/tasks/point/DropPinTaskFragmentTest.kt
@@ -22,6 +22,7 @@ import org.groundplatform.android.model.geometry.Coordinates
 import org.groundplatform.android.model.geometry.Point
 import org.groundplatform.android.model.job.Job
 import org.groundplatform.android.model.job.Style
+import org.groundplatform.android.model.map.CameraPosition
 import org.groundplatform.android.model.submission.DropPinTaskData
 import org.groundplatform.android.model.task.Task
 import org.groundplatform.android.persistence.local.LocalValueStore
@@ -29,7 +30,6 @@ import org.groundplatform.android.ui.common.ViewModelFactory
 import org.groundplatform.android.ui.datacollection.DataCollectionViewModel
 import org.groundplatform.android.ui.datacollection.components.ButtonAction
 import org.groundplatform.android.ui.datacollection.tasks.BaseTaskFragmentTest
-import org.groundplatform.android.ui.map.CameraPosition
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
